comments;move types to common
