+++ sources/xray-source/package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "axios-auth-refresh": "^3.3.6",
     "faros-airbyte-cdk": "0.0.1",
+    "faros-airbyte-common": "0.0.1",
     "typescript-memoize": "^1.1.0"
   },
   "jest": {
