+++ sources/xray-source/resources/queries/get-tests.gql
@@ -25,6 +25,8 @@ query getTests(
       steps {
         id
         action
+        data
+        result
       }
       jira(fields: ["key", "description", "summary", "labels"])
       preconditions(limit: $limit) {
